Removed double newline from send command.
If you added a newline to the send, the server would add a 2nd newline.
Now it's up to the sender to choose a newline. This required refactoring
of all commands. Also added help commands when connect to websocket.
Also inlined the home.html file so no more dependency errors.


Former-commit-id: 506a90befec4e2d0c240c20e57b6aa3c255e737d